Use random_bytes to ensure cryptographic strong random values
* Addresses insecure openssl_random_pseudo_bytes on certain PHP versions
* Add random_compat polyfill for PHP <7.0

The random_compat dependency is defined as open as possible. Since 2.0.0
openssl has been completely removed from it. This already happened
earlier in 1.3.0 and 1.3.1, however because of downstream issues openssl
was introduced again from 1.4.0 and up in the 1.x release train.
Therefor we can tolerated both 1.3.x and 2.x versions of random_compat.