Travis: add PHP 7.1, move hhvm to allowed failures
Composer installs a version of phpunit that is not compatible with
the HHVM version that Travis supports. That can be worked around
with lots of hoop jumping, but that seems out of scope for xmlseclibs.
Just allow hhvm to fail until Travis/HHVM supply a workable combo
out of the box.